refactor: colocate rule cache with grammar compiler
Keep the compiler-owned cross-grammar cache private to its only complete-type consumer and remove the unrelated compiled-grammar dependency from grammar functors.
